some fix

+++ lesson_01/pythagoras.rb
@@ -4,7 +4,7 @@
 print 'Ввведите стороны треугольника A B C через пробел: '
 abc = gets.chomp.split(" ")
 
-a, b, c = abc.map!{ |x| x.to_i }.sort!
+a, b, c = abc.map!{ |x| x.to_f }.sort!
 
 if a**2 + b**2 == c**2
   puts 'Этот треугольник прямоугольный.'
